+++ __init__.py
@@ -44,7 +44,7 @@ from .Nodes.Qwen3VL.MultiplePathsInput import MultiplePathsInput
 from .Nodes.Qwen3VL.PathSwitch import PathSwitch
 from .Nodes.ImageRotateTool import ImageRotateTool
 from .Nodes.ImageFormatConverter import ImageFormatConverter
-from .Nodes.Qwen3VL.Qwen3VLAPI import Qwen3VLAPI
+from .Nodes.Qwen3VL_API.Qwen3VLAPI import Qwen3VLAPI
 from .Nodes.Sa2VA.Sa2VAAdvanced import Sa2VAAdvanced
 from .Nodes.Sa2VA.Sa2VASegmentationPreset import Sa2VASegmentationPreset
 from .Nodes.Batch_loading_of_images import BatchLoadingOfImages
@@ -98,10 +98,10 @@ NODE_CLASS_MAPPINGS = {
     "MultiplePathsInput": MultiplePathsInput,
     "ImageRotateTool": ImageRotateTool,
     "ImageFormatConverter": ImageFormatConverter,
-    "Qwen3VLAPI": Qwen3VLAPI,
     "Sa2VAAdvanced": Sa2VAAdvanced,
     "Sa2VASegmentationPreset": Sa2VASegmentationPreset,
     "BatchLoadingOfImages": BatchLoadingOfImages,
+    "Qwen3VLAPI": Qwen3VLAPI,
 }
 
 NODE_DISPLAY_NAME_MAPPINGS = {
@@ -151,10 +151,10 @@ NODE_DISPLAY_NAME_MAPPINGS = {
     "PathSwitch": "Qwen3-VL Path Switch",
     "ImageRotateTool": "Image Rotate Tool",
     "ImageFormatConverter": "Image Format Converter",
-    "Qwen3VLAPI": "Qwen3-VL API",
     "Sa2VAAdvanced": "Sa2VA Advanced",
     "Sa2VASegmentationPreset": "Sa2VA Segmentation Preset",
     "BatchLoadingOfImages": "Batch Loading Of Images",
+    "Qwen3VLAPI": "Qwen3-VL API",
 }
 
 WEB_DIRECTORY = os.path.join(os.path.dirname(os.path.abspath(__file__)), "web")